Adapt tests to long doc ids and add coverage beyond 2^31
Update the test framework and core tests for the widened ScoreDoc/reader
APIs, and add coverage that does not require indexing billions of docs:
TestSearchBeyondMaxInt wraps tiny real segments in FilterLeafReaders that
report an inflated maxDoc so a real hit lands past 2^31; TestDocScoreHeap
checks both heap backings against a brute-force top-K; TestMaxDocsAcross-
Segments lowers the per-segment cap to exercise the writer/merge paths.
Also fixes the test framework resetting the whole-index cap after every
test, which had hidden that the >2^31 path was untested.

+++ lucene/core/src/test/org/apache/lucene/codecs/perfield/TestPerFieldDocValuesFormat.java
@@ -123,7 +123,7 @@ public class TestPerFieldDocValuesFormat extends BaseDocValuesFormatTestCase {
     StoredFields storedFields = isearcher.storedFields();
     // Iterate through the results:
     for (int i = 0; i < hits.scoreDocs.length; i++) {
-      int hitDocID = hits.scoreDocs[i].doc;
+      int hitDocID = Math.toIntExact(hits.scoreDocs[i].doc);
       Document hitDoc = storedFields.document(hitDocID);
       assertEquals(text, hitDoc.get("fieldname"));
       assert ireader.leaves().size() == 1;
