+++ lucene/core/src/test/org/apache/lucene/index/TestIndexingSequenceNumbers.java
@@ -483,7 +483,7 @@ public class TestIndexingSequenceNumbers extends LuceneTestCase {
         // We pre-add all ids up front:
         assert expectedThreadIDs[id] != -1;
         assertEquals(1, hits.totalHits.value());
-        int hitDoc = hits.scoreDocs[0].doc;
+        int hitDoc = Math.toIntExact(hits.scoreDocs[0].doc);
         assertEquals(hitDoc, docValues.advance(hitDoc));
         int actualThreadID = (int) docValues.longValue();
         if (expectedThreadIDs[id] != actualThreadID) {
